requirements.txt
google-genai
google-generativeai
streamlit
python-dotenv